# docs/swagger.yaml
definitions:
  main.requestAddUser:
    properties:
      passportNumber:
        type: string
    type: object
  main.requestUpdateUser:
    properties:
      address:
        type: string
      name:
        type: string
      passportNumber:
        type: integer
      passportSerie:
        type: integer
      patronymic:
        type: string
      surname:
        type: string
    type: object
  main.responseAddUser:
    properties:
      user:
        $ref: '#/definitions/model.User'
    type: object
  main.responseShowUsers:
    properties:
      users:
        items:
          $ref: '#/definitions/model.User'
        type: array
    type: object
  main.responseUpdateUser:
    properties:
      user:
        $ref: '#/definitions/model.User'
    type: object
  model.User:
    properties:
      address:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      passportSerie:
        type: integer
      passwortNumber:
        type: integer
      patronymic:
        type: string
      surname:
        type: string
      updatedAt:
        type: string
    type: object
  validator.Validator:
    properties:
      errors:
        items:
          type: string
        type: array
      fieldErrors:
        additionalProperties:
          type: string
        type: object
    type: object
info:
  contact: {}
paths:
  /sessions/{userId}/{taskId}:
    post:
      description: Start session
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad request input
          schema:
            type: object
        "409":
          description: Session already exists
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      summary: Start Session
      tags:
      - sessions
  /status:
    get:
      consumes:
      - application/json
      description: Check if the server is up and running
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Server Status
      tags:
      - api
  /users:
    get:
      description: Show all users by filters with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        minimum: 1
        name: pageSize
        type: integer
      - description: User name
        in: query
        name: name
        type: string
      - description: User surname
        in: query
        name: surname
        type: string
      - description: User patronymic
        in: query
        name: patronymic
        type: string
      - description: User address
        in: query
        name: address
        type: string
      - description: User passport serie
        in: query
        name: passportSerie
        type: integer
      - description: User passport number
        in: query
        name: passportNumber
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.responseShowUsers'
        "400":
          description: Bad request
          schema:
            type: object
        "422":
          description: Invalid input data
          schema:
            $ref: '#/definitions/validator.Validator'
        "500":
          description: Internal server error
          schema:
            type: object
      summary: Show Users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Add new user
      parameters:
      - description: Passport serie and number
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/main.requestAddUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.responseAddUser'
        "400":
          description: Bad request input
          schema:
            type: object
        "409":
          description: User already exists
          schema:
            type: object
        "422":
          description: Invalid input data
          schema:
            $ref: '#/definitions/validator.Validator'
        "500":
          description: Internal server error
          schema:
            type: object
      summary: Add User
      tags:
      - users
  /users/{userId}:
    delete:
      description: Delete user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad request input
          schema:
            type: object
        "404":
          description: User not found
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      summary: Delete User
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      - description: New user data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/main.requestUpdateUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.responseUpdateUser'
        "400":
          description: Bad request
          schema:
            type: object
        "404":
          description: User not found
          schema:
            type: object
        "409":
          description: User already exists
          schema:
            type: object
        "422":
          description: Invalid input data
          schema:
            $ref: '#/definitions/validator.Validator'
        "500":
          description: Internal server error
          schema:
            type: object
      summary: Update user
      tags:
      - users
swagger: "2.0"